added onclick call back
